Выполняет задание "Модуляция"
